# [Go] Fix performance issues with expired cache (decrypt) - issue #385