--- a/go/appencryption/cache.go
+++ b/go/appencryption/cache.go
@@ -96,9 +96,6 @@ func isReloadRequired(entry cacheEntry, checkInterval time.Duration) bool {
 // is not present in the cache it will retrieve the key using the provided keyLoader
 // and store the key if an error is not returned.
 func (c *keyCache) GetOrLoad(id KeyMeta, loader keyLoader) (*internal.CryptoKey, error) {
-	c.rw.Lock()
-	defer c.rw.Unlock()
-
 	if k, ok := c.get(id); ok {
 		return k, nil
 	}
@@ -112,6 +109,8 @@ func (c *keyCache) GetOrLoad(id KeyMeta, loader keyLoader) (*internal.CryptoKey,
 // The second return value indicates the successful retrieval of a
 // fresh key.
 func (c *keyCache) get(id KeyMeta) (*internal.CryptoKey, bool) {
+	c.rw.RLock()
+	defer c.rw.RUnlock()
 	key := cacheKey(id.ID, id.Created)
 
 	if e, ok := c.read(key); ok && !isReloadRequired(e, c.policy.RevokeCheckInterval) {
@@ -127,6 +126,9 @@ func (c *keyCache) get(id KeyMeta) (*internal.CryptoKey, bool) {
 // load maintains the latest entry for each distinct ID which can be accessed using
 // id.Created == 0.
 func (c *keyCache) load(id KeyMeta, loader keyLoader) (*internal.CryptoKey, error) {
+	c.rw.Lock()
+	defer c.rw.Unlock()
+
 	key := cacheKey(id.ID, id.Created)
 
 	k, err := loader.Load()
@@ -140,6 +142,7 @@ func (c *keyCache) load(id KeyMeta, loader keyLoader) (*internal.CryptoKey, erro
 		// we just loaded since we don't need it
 		e.key.SetRevoked(k.Revoked())
 		e.loadedAt = time.Now()
+		c.write(key, e)
 
 		k.Close()
 	} else {