This should fix issue #385 
It appears that when the cache item was expired and then reloaded, it was never written back to the cache.  This caused the item to get loaded every time after RevokeCheckInterval  expired.

There is a small fix for writing back to the cache, and then I also moved the locks around to load/get.   This should allow for better performance on reads with multiple readers.  I also ran "go mod tidy".  I added a benchmark test and normal test so we can test this in the future. 
To run the benchmark, try something like this:
`go test -run="notthis" -bench='^BenchmarkKeyCache_GetOrLoad_MultipleThreadsRead_NeedReloadKey$' -v -cpu=12 -cpuprofile=cpu.out -benchtime=60s 
`
I noticed significant improvements after making the "write" fix.  After making the "write" fix + "lock changes", I found a slight improvement as well.

CPU profiles:

- Original: `BenchmarkKeyCache_GetOrLoad_MultipleThreadsRead_NeedReloadKey-12           12675           5642228 ns/op`

- With Write Fix Only: `BenchmarkKeyCache_GetOrLoad_MultipleThreadsRead_NeedReloadKey-12        98069263              1094 ns/op`

- With Write Fix And Lock Changes: `BenchmarkKeyCache_GetOrLoad_MultipleThreadsRead_NeedReloadKey-12        100000000              803.6 ns/op`

I have attached the [cpuprofiles.zip](https://github.com/godaddy/asherah/files/6849304/cpuprofiles.zip)
 for all three scenarios.

1. Original
2. Write fix only
3. Write fix with lock changes

Here we are comparing the original with the write fix only:
`go tool pprof -nodefraction=0 -http=:8086 -normalize -diff_base cpu_original.out cpu_write_fix.out 
`![cpu_diff_write_fix](https://user-images.githubusercontent.com/1618767/126345466-e93dfbd0-ceec-40fb-952a-23bf6bc7a275.png)


Here is comparing write fix with write fix and lock changes : 
`go tool pprof -nodefraction=0 -http=:8086 -normalize -diff_base cpu_write_fix.out cpu_write_fix_with_lock_changes.out 
`![image](https://user-images.githubusercontent.com/1618767/126343817-5a78aa73-d9fd-413c-81fa-b83d19a5fee6.png)

We ran a small canary test in prod.  With the RevokeCheckInterval set to 5 minutes, we saw much better performance with this fix.
![Datafabric_DRR_metrics_after_fix](https://user-images.githubusercontent.com/1618767/126345724-95f1935a-77af-4136-baa8-84353d5b1dec.png)