@keperry 2021-07-20
Note:  Modified locks to use RLock and Lock on GetOrLoad().  This should get us both quick reads and performant writes.  I also added an early exit for the heavy "Lock()" - that should prevent several queued up RW locks from all performing loads.   Here are the profiles for both scenarios:  
[cpuprofiles_take2.zip](https://github.com/godaddy/asherah/files/6851021/cpuprofiles_take2.zip)


Here are some results from my machine - the performance looks a lot better with the lock changes:
- Write Fix only: 
`BenchmarkKeyCache_GetOrLoad_MultipleThreadsRead_NeedReloadKey-12        129446646              585.7 ns/op`
- Write Fix with Lock Improvements:
`BenchmarkKeyCache_GetOrLoad_MultipleThreadsRead_NeedReloadKey-12        551315848              177.4 ns/op
`
![image](https://user-images.githubusercontent.com/1618767/126390859-952386b5-054b-4038-bcf5-c083651d91d9.png)